Move DataViewer class with state handling to glue.viewers.common
